Extract JsonPointer to common.py
